Restructure project
This commit restructures the project to match the pattern described in the original article, where each component serves as a different gradle module. This will help in the future if we would like to have multiple implementations of different frameworks using the same business logic, without the need to duplicate the codebase.