Add initial models
This commit introduces the foundational models for the application, including `Cart`, `CartLineItem`, `Product` and `Money`. These models are essential for the core functionalities such as adding product to a cart of the user, getting the current cart, and calculating the price of a cart. Each model is accompanied based on behavioral of it rather than slim data model