Add Konsist to the project
This commit add Konsist to the project to ensure consistency between the different components. Moreover, we're using it to ensure architecture enforcement for the hexagonal architecture as a second layer of protection (above gradle packages).

This commit also fixes all the issues that have been discovered while running the new tests